+++ Re/ScissorRect.h
@@ -7,7 +7,7 @@ private:
 	D3D12_RECT scissorRect{};
 public:
 	ScissorRect();
-	void Init(const int left, const int width, const int top, const int height);
+	void Init(const float left, const float width, const float top, const float height);
 	void Update(ID3D12GraphicsCommandList* cmdList);
 };
